refactor: Add authorization middleware for instructors
This commit adds a new middleware function called `authorizeInstructor` to the `isAuthorized.js` file in the `middlewares` directory. The `authorizeInstructor` function checks if the user's role is not equal to 'instructor' and returns a 403 Forbidden response if it is not. This middleware is intended to be used to restrict certain actions to only instructors.