refactor: Update available modules query to filter by user ID
The available modules query in the availablemodules.js controller was updated to filter the modules by the user ID. This change ensures that only modules created by the specified user are returned, improving the accuracy of the available modules list.

Note: This message is based on the provided code changes and recent commits.